Stations: aggregated departures, children and routes; pattern fallbacks; plan endpoint names
- /stops/{station} and /departures resolve parent stations via OTP station(id);
  departures are deduped by tripId and sorted; children/routes come from
  parent_station when OTP has none. Stop names are trimmed.
- Route patterns: headsign falls back to the last stop, directionId null when unknown.
- nearby distanceMeters is an integer; geocode type union documents 'place'.
- /plan accepts fromName/toName and otherwise reverse-geocodes within 1.5 s.

+++ app/models.py
@@ -165,7 +165,7 @@ class Stop(Out):
 
 
 class NearbyStop(Stop):
-    distance_meters: float
+    distance_meters: int
 
 
 class NearbyResponse(Out):
